@@ -47,40 +47,40 @@
     "postpublish": "npm run clean && npm run build:docsite && git add docs && git commit -m 'Building docs' &&  git push && git push --tags && npm run clean"
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^5.0.11",
-    "@storybook/addon-knobs": "^5.0.11",
-    "@storybook/react": "^5.0.11",
-    "@types/enzyme": "^3.9.1",
+    "@storybook/addon-actions": "^5.1.11",
+    "@storybook/addon-knobs": "^5.1.11",
+    "@storybook/react": "^5.1.11",
+    "@types/enzyme": "^3.10.3",
     "@types/enzyme-adapter-react-16": "^1.0.5",
-    "@types/jest": "^24.0.9",
-    "@types/node": "^10.12.18",
-    "@types/react": "^16.8.3",
-    "@types/react-dom": "^16.0.11",
-    "@types/react-router": "^4.4.4",
-    "@types/react-router-dom": "^4.3.1",
-    "@types/storybook__addon-actions": "^3.4.2",
-    "@types/storybook__addon-knobs": "^5.0.0",
-    "@types/storybook__react": "^4.0.1",
-    "babel-loader": "8.0.5",
-    "copyfiles": "^2.1.0",
+    "@types/jest": "^24.0.18",
+    "@types/node": "^12.7.4",
+    "@types/react": "^16.9.2",
+    "@types/react-dom": "^16.9.0",
+    "@types/react-router": "^5.0.3",
+    "@types/react-router-dom": "^4.3.5",
+    "@types/storybook__addon-actions": "^3.4.3",
+    "@types/storybook__addon-knobs": "^5.0.3",
+    "@types/storybook__react": "^4.0.2",
+    "babel-loader": "^8.0.6",
+    "copyfiles": "^2.1.1",
     "documentalist": "^1.6.0",
-    "enzyme": "^3.9.0",
-    "enzyme-adapter-react-16": "^1.12.1",
-    "generate-changelog": "^1.7.1",
-    "lerna": "^3.14.1",
-    "module-alias": "^2.2.0",
-    "react": "^16.8.1",
-    "react-app-rewire-hot-loader": "^1.0.3",
-    "react-app-rewired": "^2.0.2",
-    "react-dom": "^16.8.1",
-    "react-ga": "^2.5.7",
-    "react-hot-loader": "^4.6.3",
-    "react-live": "^2.0.0",
-    "react-markdown": "^4.0.6",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "^3.0.0",
-    "typescript": "^3.3.3333",
-    "webpack": "4.29.6"
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.14.0",
+    "generate-changelog": "^1.8.0",
+    "lerna": "^3.16.4",
+    "module-alias": "^2.2.1",
+    "react": "^16.9.0",
+    "react-app-rewire-hot-loader": "^2.0.1",
+    "react-app-rewired": "^2.1.3",
+    "react-dom": "^16.9.0",
+    "react-ga": "^2.6.0",
+    "react-hot-loader": "^4.12.12",
+    "react-live": "^2.2.0",
+    "react-markdown": "^4.2.2",
+    "react-router-dom": "^5.0.1",
+    "react-scripts": "^3.1.1",
+    "typescript": "^3.5.3",
+    "webpack": "^4.39.1"
   },
   "eslintIgnore": [
     "node_modules"
