@@ -2,14 +2,16 @@ import * as React from 'react';
 
 export type TabBarContextT = {
   onTabInteraction: (evt: any) => void;
-  registerTab: (tab: any) => {};
-  unregisterTab: (tab: any) => {};
+  registerTab: (tab: any) => void;
+  unregisterTab: (tab: any) => void;
+  indicatorTransition: 'slide' | 'fade';
 };
 
-export const TabBarContext = React.createContext({
+export const TabBarContext = React.createContext<TabBarContextT>({
   onTabInteraction: (evt: any) => {},
   registerTab: (tab: any) => {},
-  unregisterTab: (tab: any) => {}
+  unregisterTab: (tab: any) => {},
+  indicatorTransition: 'slide'
 });
 
 export const withTabBarContext = () => <P extends {}>(
