@@ -1,3 +1,4 @@
+import * as RMWC from '@rmwc/types';
 import * as React from 'react';
 
 import {
@@ -5,17 +6,25 @@ import {
   MDCTabIndicatorFoundation,
   MDCSlidingTabIndicatorFoundation
 } from '@material/tab-indicator';
-import { FoundationComponent } from '@rmwc/base';
+import { FoundationComponent, classNames } from '@rmwc/base';
+import { Icon } from '@rmwc/icon';
+
+export interface TabIndicatorProps {
+  /** Use an icon instead of an underline for the tab */
+  icon?: RMWC.IconPropT;
+  /** The transition to use */
+  transition?: 'slide' | 'fade';
+}
 
 export class TabIndicator extends FoundationComponent<
-  MDCSlidingTabIndicatorFoundation,
-  {}
+  MDCTabIndicatorFoundation,
+  TabIndicatorProps
 > {
   private root = this.createElement('root');
   private content = this.createElement('content');
 
   getDefaultFoundation() {
-    return new MDCSlidingTabIndicatorFoundation({
+    const adapter = {
       addClass: (className: string) => {
         this.root.addClass(className);
       },
@@ -29,7 +38,13 @@ export class TabIndicator extends FoundationComponent<
       setContentStyleProperty: (prop: string, value: string) => {
         this.content.setStyle(prop, value);
       }
-    });
+    };
+
+    if (this.props.transition === 'fade') {
+      return new MDCFadingTabIndicatorFoundation(adapter);
+    }
+
+    return new MDCSlidingTabIndicatorFoundation(adapter);
   }
 
   activate(previousIndicatorClientRect?: ClientRect) {
@@ -45,12 +60,25 @@ export class TabIndicator extends FoundationComponent<
   }
 
   render() {
+    const { icon, transition } = this.props;
+    const Tag = !!icon ? Icon : 'span';
+
     return (
-      <span {...this.root.props({ className: 'mdc-tab-indicator' })}>
-        <span
+      <span
+        {...this.root.props({
+          className: classNames('mdc-tab-indicator', {
+            'mdc-tab-indicator--fade': transition === 'fade'
+          })
+        })}
+      >
+        <Tag
+          aria-hidden="true"
+          icon={icon}
           ref={this.content.setRef}
           {...this.content.props({})}
-          className="mdc-tab-indicator__content mdc-tab-indicator__content--underline"
+          className={`mdc-tab-indicator__content mdc-tab-indicator__content--${
+            icon ? 'icon' : 'underline'
+          }`}
         />
       </span>
     );
