@@ -1,10 +1,11 @@
-import React from 'react';
+import React, { Children } from 'react';
 
-import { Docs, DocsExample, DocProps } from '@doc-utils';
+import { Docs, DocsP, DocsExample, DocProps, DocsSubtitle } from '@doc-utils';
 import propsSrc from './generated-props.json';
 import examples from './generated-examples.json';
 
 import { Typography } from '.';
+import { RMWCProvider } from '../provider';
 
 export default function() {
   return (
@@ -16,6 +17,7 @@ export default function() {
       docsLink="https://material.io/develop/web/components/typography/"
       examples={examples}
     >
+      <DocsSubtitle>Basic Usage</DocsSubtitle>
       <DocsExample>
         <>
           <Typography use="headline1">headline1</Typography>
@@ -32,6 +34,42 @@ export default function() {
           <Typography use="button">button</Typography>
           <Typography use="overline">overline</Typography>
         </>
+      </DocsExample>
+
+      <DocsExample label="Specify tag">
+        <Typography use="headline6" tag="h3">
+          headline6
+        </Typography>
+      </DocsExample>
+
+      <DocsSubtitle>Provider Setup</DocsSubtitle>
+      <DocsP>
+        You can globally map specify type classes to HTML tags for semantics, or
+        specify a default to be used instead of `span`.
+      </DocsP>
+
+      <DocsExample label="Specify tag">
+        <RMWCProvider
+          typography={{
+            /** Make all Typography components default to <div>  */
+            defaultTag: 'div',
+            /** Make headline5 <h5>  */
+            headline5: 'h5',
+            /** Make body2 <p>  */
+            body2: 'p',
+            /** Use your own component  */
+            body1: ({ children, ...rest }) => (
+              <span>
+                <b>{children}!!!</b>
+              </span>
+            )
+          }}
+        >
+          <Typography use="headline6">Rendered default `div`</Typography>
+          <Typography use="headline5">Rendered with `h5`</Typography>
+          <Typography use="body2">Rendered with `p`</Typography>
+          <Typography use="body1">Custom rendering</Typography>
+        </RMWCProvider>
       </DocsExample>
 
       <DocProps src={propsSrc} components={[Typography]} />
